requirements.txt
# ─────────────────────────────────────────────
# Core Streamlit & LangChain
# ─────────────────────────────────────────────
streamlit>=1.32.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# ─────────────────────────────────────────────
# LLM Providers
# ─────────────────────────────────────────────
langchain-groq>=0.1.6
langchain-openai>=0.1.8
langchain-google-genai>=1.0.6

# ─────────────────────────────────────────────
# Embeddings
# ─────────────────────────────────────────────
langchain-huggingface>=0.0.3
sentence-transformers>=2.7.0
torch>=2.0.0

# ─────────────────────────────────────────────
# Vector Store
# ─────────────────────────────────────────────
faiss-cpu>=1.7.4

# ─────────────────────────────────────────────
# Document Processing
# ─────────────────────────────────────────────
pdfplumber>=0.10.3
PyPDF2>=3.0.1

# ─────────────────────────────────────────────
# Web Search
# ─────────────────────────────────────────────
tavily-python>=0.3.3
duckduckgo-search>=6.1.0

# ─────────────────────────────────────────────
# Utilities
# ─────────────────────────────────────────────
python-dotenv>=1.0.0
